pyproject.toml
[tool.poetry]
name = "drf_shoestore"
version = "0.1.0"
description = "Kenzie ReST Framework:  APIs in a Nutshell assessment project."
authors = ["Jackson Nathaniel Detke <jndetke2150@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.1"
djangorestframework = "^3.12.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
